haproxy-config-generator-app/requirements.txt
jinja2==2.7.3
PyYAML==3.11